Get auto field information per user from config file
If there are multiple authors in the blog, some fields can be related to
the account. This commit provides this options.

Adding accounts fields in the config.yml files like this following:

accounts:
  - email: myid@email.com
    auto_fields:
        author: john

This configuration makes the 'author' field in frontmatter be filled
with 'john' automatically when myid@email.com has logged in and the
field is empty.